🐛fix : #72 Header 컴포넌트 2차 분할 작업
